+++ LPTestTarget/FirstViewController.h
@@ -1,0 +1,7 @@
+#import <UIKit/UIKit.h>
+
+@interface FirstViewController : UIViewController
+
+@property (weak, nonatomic) IBOutlet UIWebView *webView;
+
+@end