+++ calabash-ios-server-tests/LPJSONUtilsTest.m
@@ -2,12 +2,8 @@
 #warning This file must be compiled with ARC. Use -fobjc-arc flag (or convert project to ARC).
 #endif
 
-#import <UIKit/UIKit.h>
-#import <XCTest/XCTest.h>
 #import "LPJSONUtils.h"
-#import <OCMock/OCMock.h>
 #import "LPTouchUtils.h"
-#import "objc/runtime.h"
 
 @interface LPXCTestSliderWithText : UISlider
 
@@ -17,14 +13,21 @@
 
 @implementation LPXCTestSliderWithText
 
-@synthesize text;
+@end
+
+@interface LPObjectRetunsFrame : NSObject
+
+@property(copy, nonatomic) NSString *frame;
+
+@end
+
+@implementation LPObjectRetunsFrame
 
 @end
 
 // TODO:  UIBarButton
 // TODO:  UINavigationBarButton
 // TODO:  UITabBarButton
-// TODO:  Test with an Application Target so we can verify view frame transforms.
 @interface LPJSONUtils (LPXCTTEST)
 
 + (void) dictionary:(NSMutableDictionary *) dictionary
@@ -32,7 +35,8 @@
          whenTarget:(id) target
          respondsTo:(SEL) selector;
 
-+ (NSMutableDictionary *) jsonifyView:(UIView *) view;
++ (NSMutableDictionary *) dictionaryByEncodingView:(id) view;
++ (void) insertHitPointIntoMutableDictionary:(NSMutableDictionary *) dictionary;
 
 @end
 
@@ -48,17 +52,6 @@
 
 - (void)tearDown {
   [super tearDown];
-}
-
-#pragma mark - Mocking
-
-- (id) mockForLPTouchUtilsWindowForView:(id) view {
-  CGRect frame = [[UIScreen mainScreen] applicationFrame];
-  UIWindow *window = [[UIWindow alloc] initWithFrame:frame];
-  [window addSubview:view];
-  id mock = [OCMockObject mockForClass:[LPTouchUtils class]];
-  [[[mock stub] andReturn:window] windowForView:view];
-  return mock;
 }
 
 #pragma mark - dictionary:setObject:forKey:whenTarget:respondsTo:
@@ -98,7 +91,39 @@
   XCTAssertEqualObjects(dict[@"key"], nil);
 }
 
-#pragma mark - jsonifyView: setting value
+#pragma mark - dictionaryByEncodingView:  called on an object that is not a UIView
+
+- (void) testJsonifyViewPassedAString {
+  NSString *string = @"string";
+
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:string];
+
+  XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
+  XCTAssertEqualObjects(dict[@"class"], NSStringFromClass([string class]));
+  XCTAssertEqualObjects(dict[@"description"], string);
+  XCTAssertEqualObjects(dict[@"id"], [NSNull null]);
+  XCTAssertEqualObjects(dict[@"label"], [NSNull null]);
+  XCTAssertEqualObjects(dict[@"value"], [NSNull null]);
+  XCTAssertEqual([dict count], 6);
+}
+
+- (void) testJsonifyViewPassedAnNonViewObjectThatRespondsToFrame {
+  LPObjectRetunsFrame *framer = [LPObjectRetunsFrame new];
+  framer.frame = @"a frame";
+
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:framer];
+
+  XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
+  XCTAssertEqualObjects(dict[@"class"],
+                        NSStringFromClass([LPObjectRetunsFrame class]));
+  XCTAssertEqualObjects(dict[@"description"], [framer description]);
+  XCTAssertEqualObjects(dict[@"id"], [NSNull null]);
+  XCTAssertEqualObjects(dict[@"label"], [NSNull null]);
+  XCTAssertEqualObjects(dict[@"value"], [NSNull null]);
+  XCTAssertEqual([dict count], 6);
+}
+
+#pragma mark - dictionaryByEncodingView: setting 'value' key
 
 // Is UISlider the only UIView with value selector?
 - (void) testJsonfiyViewValueSettingHasValueTextAccessibilityValue {
@@ -111,9 +136,10 @@
   float expected = 0.5f;
   view.value = expected;
 
-  view.text = @"TEXT!";
+  NSString *text = @"TEXT!";
+  view.text = text;
 
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -126,9 +152,9 @@
   XCTAssertEqualObjects(dict[@"frame"][@"width"], @(CGRectGetWidth([view frame])));
   XCTAssertEqualObjects(dict[@"frame"][@"height"], @(CGRectGetHeight([view frame])));
   XCTAssertEqualObjects(dict[@"id"], [NSNull null]);
-  XCTAssertEqualObjects(dict[@"label"], [NSNull null]);
+  XCTAssertEqualObjects(dict[@"label"], text);
   XCTAssertEqualObjects(dict[@"selected"], @(0));
-  XCTAssertEqualObjects(dict[@"text"], @"TEXT!");
+  XCTAssertEqualObjects(dict[@"text"], text);
   XCTAssertEqualObjects(dict[@"value"], @(expected));
   XCTAssertEqualObjects(dict[@"visible"], @(1));
   XCTAssertEqual([dict count], 12);
@@ -139,9 +165,10 @@
 
   UITextField *view = [[UITextField alloc] initWithFrame:frame];
   view.accessibilityValue = @"ACCESSIBILITY VALUE!";
-  view.text = @"TEXT!";
+  NSString *text = @"TEXT!";
+  view.text = text;
 
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -156,8 +183,8 @@
   XCTAssertEqualObjects(dict[@"id"], [NSNull null]);
   XCTAssertEqualObjects(dict[@"label"], [NSNull null]);
   XCTAssertEqualObjects(dict[@"selected"], @(0));
-  XCTAssertEqualObjects(dict[@"text"], @"TEXT!");
-  XCTAssertEqualObjects(dict[@"value"], @"TEXT!");
+  XCTAssertEqualObjects(dict[@"text"], text);
+  XCTAssertEqualObjects(dict[@"value"], text);
   XCTAssertEqualObjects(dict[@"visible"], @(1));
   XCTAssertEqual([dict count], 12);
 }
@@ -166,9 +193,10 @@
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
 
   UIView *view = [[UIView alloc] initWithFrame:frame];
-  view.accessibilityValue = @"ACCESSIBILITY VALUE!";
+  NSString *value = @"ACCESSIBILITY VALUE!";
+  view.accessibilityValue = value;
 
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -182,20 +210,24 @@
   XCTAssertEqualObjects(dict[@"frame"][@"height"], @(CGRectGetHeight([view frame])));
   XCTAssertEqualObjects(dict[@"id"], [NSNull null]);
   XCTAssertEqualObjects(dict[@"label"], [NSNull null]);
-  XCTAssertEqualObjects(dict[@"value"], @"ACCESSIBILITY VALUE!");
+  XCTAssertEqualObjects(dict[@"value"], value);
   XCTAssertEqualObjects(dict[@"visible"], @(1));
   XCTAssertEqual([dict count], 10);
 }
 
-#pragma mark - jsonifyView: when view is subview of a window
+#pragma mark - dictionaryByEncodingView: when view is subview of a window
 
-- (void) testJsonifyViewUIViewWithWindow {
+- (void) testJsonifyViewUIViewWithMockedWindow {
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UIView *view = [[UIView alloc] initWithFrame:frame];
-  id mock = [self mockForLPTouchUtilsWindowForView:view];
 
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
-  NSLog(@"%@", dict);
+  CGRect applicationFrame = [[UIScreen mainScreen] applicationFrame];
+  UIWindow *window = [[UIWindow alloc] initWithFrame:applicationFrame];
+  [window addSubview:view];
+  id mock = [OCMockObject mockForClass:[LPTouchUtils class]];
+  [[[mock stub] andReturn:window] windowForView:view];
+
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -214,12 +246,12 @@
   XCTAssertEqualObjects(dict[@"frame"][@"width"], @(CGRectGetWidth([view frame])));
   XCTAssertEqualObjects(dict[@"frame"][@"height"], @(CGRectGetHeight([view frame])));
   XCTAssertEqual(((NSDictionary *)[dict objectForKey:@"rect"]).count, 6);
-  XCTAssertEqualObjects(dict[@"rect"][@"x"], @(0));
-  XCTAssertEqualObjects(dict[@"rect"][@"y"], @(0));
-  XCTAssertEqualObjects(dict[@"rect"][@"width"],@(0));
-  XCTAssertEqualObjects(dict[@"rect"][@"height"], @(0));
-  XCTAssertEqualObjects(dict[@"rect"][@"center_x"], @(0));
-  XCTAssertEqualObjects(dict[@"rect"][@"center_y"], @(0));
+  XCTAssertEqualObjects(dict[@"rect"][@"x"], @(CGRectGetMinX([view frame])));
+  XCTAssertEqualObjects(dict[@"rect"][@"y"], @(CGRectGetMinY([view frame])));
+  XCTAssertEqualObjects(dict[@"rect"][@"width"], @(CGRectGetWidth([view frame])));
+  XCTAssertEqualObjects(dict[@"rect"][@"height"], @(CGRectGetHeight([view frame])));
+  XCTAssertEqualObjects(dict[@"rect"][@"center_x"], @(CGRectGetMidX([view frame])));
+  XCTAssertEqualObjects(dict[@"rect"][@"center_y"], @(106.75));
   XCTAssertEqualObjects(dict[@"value"], [NSNull null]);
   XCTAssertEqualObjects(dict[@"visible"], @(0));
   XCTAssertEqual([dict count], 11);
@@ -227,12 +259,51 @@
   [mock verify];
 }
 
-#pragma mark - jsonifyView:
+- (void) testJsonifyViewUIViewWithApplicationWindow {
+  CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
+  UIView *view = [[UIView alloc] initWithFrame:frame];
+
+  UIApplication *app = [UIApplication sharedApplication];
+  UIWindow *window = [[app delegate] window];
+  XCTAssertNotNil(window);
+  [window addSubview:view];
+
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
+
+  XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
+  XCTAssertEqualObjects(dict[@"alpha"], @(1));
+  XCTAssertEqualObjects(dict[@"class"], NSStringFromClass([view class]));
+  XCTAssertEqualObjects(dict[@"description"], [view description]);
+  XCTAssertEqualObjects(dict[@"enabled"], @(1));
+  XCTAssertEqualObjects(dict[@"frame"][@"x"], @(CGRectGetMinX([view frame])));
+  XCTAssertEqualObjects(dict[@"frame"][@"y"], @(CGRectGetMinY([view frame])));
+  XCTAssertEqualObjects(dict[@"frame"][@"width"], @(CGRectGetWidth([view frame])));
+  XCTAssertEqualObjects(dict[@"frame"][@"height"], @(CGRectGetHeight([view frame])));
+  XCTAssertEqual(((NSDictionary *)[dict objectForKey:@"frame"]).count, 4);
+  XCTAssertEqualObjects(dict[@"id"], [NSNull null]);
+  XCTAssertEqualObjects(dict[@"label"], [NSNull null]);
+  XCTAssertEqualObjects(dict[@"frame"][@"x"], @(CGRectGetMinX([view frame])));
+  XCTAssertEqualObjects(dict[@"frame"][@"y"], @(CGRectGetMinY([view frame])));
+  XCTAssertEqualObjects(dict[@"frame"][@"width"], @(CGRectGetWidth([view frame])));
+  XCTAssertEqualObjects(dict[@"frame"][@"height"], @(CGRectGetHeight([view frame])));
+  XCTAssertEqual(((NSDictionary *)[dict objectForKey:@"rect"]).count, 6);
+  XCTAssertEqualObjects(dict[@"rect"][@"x"], @(CGRectGetMinX([view frame])));
+  XCTAssertEqualObjects(dict[@"rect"][@"y"], @(CGRectGetMinY([view frame])));
+  XCTAssertEqualObjects(dict[@"rect"][@"width"], @(CGRectGetWidth([view frame])));
+  XCTAssertEqualObjects(dict[@"rect"][@"height"], @(CGRectGetHeight([view frame])));
+  XCTAssertEqualObjects(dict[@"rect"][@"center_x"], @(CGRectGetMidX([view frame])));
+  XCTAssertEqualObjects(dict[@"rect"][@"center_y"], @(CGRectGetMidY([view frame])));
+  XCTAssertEqualObjects(dict[@"value"], [NSNull null]);
+  XCTAssertEqualObjects(dict[@"visible"], @(1));
+  XCTAssertEqual([dict count], 11);
+}
+
+#pragma mark - dictionaryByEncodingView:
 
 - (void) testJsonifyViewUIView {
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UIView *view = [[UIView alloc] initWithFrame:frame];
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -254,7 +325,7 @@
 - (void) testJsonifyViewUIControl {
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UIControl *view = [[UIControl alloc] initWithFrame:frame];
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -277,7 +348,7 @@
 - (void) testJsonifyViewUITextField {
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UITextField *view = [[UITextField alloc] initWithFrame:frame];
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -301,9 +372,9 @@
 - (void) testJsonifyViewUITextView {
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UITextView *view = [[UITextView alloc] initWithFrame:frame];
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
-  XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
+  XCTAssertEqualObjects(dict[@"accessibilityElement"], @(1));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
   XCTAssertEqualObjects(dict[@"class"], NSStringFromClass([view class]));
   XCTAssertEqualObjects(dict[@"description"], [view description]);
@@ -324,7 +395,7 @@
 - (void) testJsonifyViewUISlider {
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UISlider *view = [[UISlider alloc] initWithFrame:frame];
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -347,9 +418,9 @@
 - (void) testJsonifyViewUISwitch {
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UISwitch *view = [[UISwitch alloc] initWithFrame:frame];
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
-  XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
+  XCTAssertEqualObjects(dict[@"accessibilityElement"], @(1));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
   XCTAssertEqualObjects(dict[@"class"], NSStringFromClass([view class]));
   XCTAssertEqualObjects(dict[@"description"], [view description]);
@@ -369,9 +440,9 @@
 - (void) testJsonifyViewUIButton {
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UIButton *view = [[UIButton alloc] initWithFrame:frame];
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
-  XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
+  XCTAssertEqualObjects(dict[@"accessibilityElement"], @(1));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
   XCTAssertEqualObjects(dict[@"class"], NSStringFromClass([view class]));
   XCTAssertEqualObjects(dict[@"description"], [view description]);
@@ -390,7 +461,7 @@
 - (void) testJsonifyViewUIScrollView {
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UIScrollView *view = [[UIScrollView alloc] initWithFrame:frame];
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -411,9 +482,9 @@
 - (void) testJsonifyViewUITableView {
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UITableView *view = [[UITableView alloc] initWithFrame:frame];
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
-  XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
+  XCTAssertEqualObjects(dict[@"accessibilityElement"], @(1));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
   XCTAssertEqualObjects(dict[@"class"], NSStringFromClass([view class]));
   XCTAssertEqualObjects(dict[@"description"], [view description]);
@@ -424,7 +495,7 @@
   XCTAssertEqualObjects(dict[@"frame"][@"width"], @(CGRectGetWidth([view frame])));
   XCTAssertEqualObjects(dict[@"frame"][@"height"], @(CGRectGetHeight([view frame])));
   XCTAssertEqualObjects(dict[@"id"], [NSNull null]);
-  XCTAssertEqualObjects(dict[@"label"], [NSNull null]);
+  XCTAssertEqualObjects(dict[@"label"], @"Empty list");
   XCTAssertEqualObjects(dict[@"visible"], @(1));
   XCTAssertEqual([dict count], 10);
 }
@@ -434,7 +505,7 @@
                        initWithStyle:UITableViewCellStyleDefault
                        reuseIdentifier:@"identifier"];
 
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -459,7 +530,7 @@
   UICollectionViewFlowLayout *layout = [UICollectionViewFlowLayout new];
   UICollectionView *view = [[UICollectionView alloc] initWithFrame:frame
                                               collectionViewLayout:layout];
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -481,7 +552,7 @@
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UICollectionViewCell *view = [[UICollectionViewCell alloc] initWithFrame:frame];
 
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -504,7 +575,7 @@
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UISegmentedControl *view = [[UISegmentedControl alloc] initWithFrame:frame];
 
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -527,7 +598,7 @@
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UIPickerView *view = [[UIPickerView alloc] initWithFrame:frame];
 
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -550,7 +621,7 @@
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UIDatePicker *view = [[UIDatePicker alloc] initWithFrame:frame];
 
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -574,7 +645,7 @@
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UITabBar *view = [[UITabBar alloc] initWithFrame:frame];
 
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -597,7 +668,7 @@
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UINavigationBar *view = [[UINavigationBar alloc] initWithFrame:frame];
 
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -620,7 +691,7 @@
   CGRect frame = CGRectMake(20, 64.5, 88, 44.5);
   UIToolbar *view = [[UIToolbar alloc] initWithFrame:frame];
 
-  NSDictionary *dict = [LPJSONUtils jsonifyView:view];
+  NSDictionary *dict = [LPJSONUtils dictionaryByEncodingView:view];
 
   XCTAssertEqualObjects(dict[@"accessibilityElement"], @(0));
   XCTAssertEqualObjects(dict[@"alpha"], @(1));
@@ -637,6 +708,54 @@
   XCTAssertEqualObjects(dict[@"value"], [NSNull null]);
   XCTAssertEqualObjects(dict[@"visible"], @(1));
   XCTAssertEqual([dict count], 10);
+}
+
+#pragma mark - insertHitPointIntoMutableDictionary:
+
+- (void) testInsertHitPointIntoMutableDictionaryArgNotMutable {
+  id dict = @{};
+  [LPJSONUtils insertHitPointIntoMutableDictionary:dict];
+  XCTAssertTrue([dict count] == 0);
+}
+
+- (void) testInsertHitPointIntoMutableDictionaryNoRectInDictionary {
+  NSMutableDictionary *dict = [@{} mutableCopy];
+  [LPJSONUtils insertHitPointIntoMutableDictionary:dict];
+  XCTAssertTrue([dict count] == 1);
+  NSDictionary *hitPoint = dict[@"hit-point"];
+  XCTAssertNotNil(hitPoint);
+  XCTAssertEqualObjects(hitPoint[@"x"], [NSNull null]);
+  XCTAssertEqualObjects(hitPoint[@"y"], [NSNull null]);
+}
+
+- (void) testInsertHitPointIntoMutableDictionaryRectHasNoCenterX {
+  NSMutableDictionary *dict = [@{@"rect" : @{@"center_y" : @(55)}} mutableCopy];
+  [LPJSONUtils insertHitPointIntoMutableDictionary:dict];
+  XCTAssertTrue([dict count] == 2);
+  NSDictionary *hitPoint = dict[@"hit-point"];
+  XCTAssertNotNil(hitPoint);
+  XCTAssertEqualObjects(hitPoint[@"x"], [NSNull null]);
+  XCTAssertEqualObjects(hitPoint[@"y"], @(55));
+}
+
+- (void) testInsertHitPointIntoMutableDictionaryRectHasNoCenterY {
+  NSMutableDictionary *dict = [@{@"rect" : @{@"center_x" : @(55)}} mutableCopy];
+  [LPJSONUtils insertHitPointIntoMutableDictionary:dict];
+  XCTAssertTrue([dict count] == 2);
+  NSDictionary *hitPoint = dict[@"hit-point"];
+  XCTAssertNotNil(hitPoint);
+  XCTAssertEqualObjects(hitPoint[@"x"], @(55));
+  XCTAssertEqualObjects(hitPoint[@"y"], [NSNull null]);
+}
+
+- (void) testInsertHitPointIntoMutableDictionary {
+  NSDictionary *rect = @{@"rect" : @{@"center_x" : @(55), @"center_y" : @(65)}};
+  NSMutableDictionary *dict = [rect mutableCopy];
+  [LPJSONUtils insertHitPointIntoMutableDictionary:dict];
+  XCTAssertTrue([dict count] == 2);
+  XCTAssertEqualObjects(dict[@"hit-point"][@"x"], @(55));
+  XCTAssertEqualObjects(dict[@"hit-point"][@"y"], @(65));
+  XCTAssertNotNil(dict[@"rect"]);
 }
 
 @end