+++ scripts/test/xctest.rb
@@ -1,16 +1,16 @@
 #!/usr/bin/env ruby
 require File.expand_path(File.join(File.dirname(__FILE__), '..', 'test-helpers'))
+require 'run_loop'
 
 working_dir = File.expand_path(File.join(File.dirname(__FILE__), '..', '..'))
 
-# See the note below about why we cannot use xcpretty.
-#xcpretty_available = `gem list xcpretty -i`.chomp == 'true'
+xcpretty_available = `gem list xcpretty -i`.chomp == 'true'
 
 XCODE_MAJOR_VERSION=`xcrun -k xcodebuild -version | tr -d "\n" | cut -c 7`.chomp
 if XCODE_MAJOR_VERSION == '5'
   target_simulator_name = 'iPhone Retina (4-inch)'
 else
-  target_simulator_name = 'iPhone 5'
+  target_simulator_name = 'iPhone 5s'
 end
 
 args =
@@ -23,23 +23,46 @@ args =
             '-scheme calabash-ios-server-tests',
             "-destination 'platform=iOS Simulator,name=#{target_simulator_name},OS=latest'",
             '-sdk iphonesimulator',
-            '-configuration Debug'
-      # See the note below about why we cannot use xcpretty.
-      # xcpretty_available ? '| xcpretty -c' : ''
+            '-configuration Debug',
+            xcpretty_available ? '| xcpretty -tc && exit ${PIPESTATUS[0]}' : ''
       ]
 
 Dir.chdir(working_dir) do
-  terminate_all_sims
-  # Would like to use xcpretty here, but there are two problems.
-  #
-  # 1. xcodebuild + test + xcpretty does _not_ return a non-zero exit code
-  #    when there is a compilation error.
-  # 2. We cannot use xcodebuild _without_ xcpretty on Travis CI; there is too
-  #    much output to stdout and the build will fail.
-  #
-  # To compensate, we don't use xcpretty and suppress xcodebuild output.
-  cmd = "xcrun xcodebuild #{args.join(' ')} > /dev/null"
-  do_system(cmd,
-            {:pass_msg => 'XCTests passed',
-             :fail_msg => 'XCTests failed'})
+
+  # It is possible that this is causing problems on Travis CI on the first run.
+  # (>_>) Get up get down Apple XCTest is a joke in my town.
+  unless travis_ci?
+    RunLoop::SimControl.terminate_all_sims
+  end
+
+  cmd = "xcrun xcodebuild #{args.join(' ')}"
+
+  on_retry = Proc.new do |_, try, elapsed_time, next_interval|
+    log_fail "XCTest: attempt #{try} failed in '#{elapsed_time}'; will retry in '#{next_interval}'"
+    RunLoop::SimControl.terminate_all_sims
+  end
+
+  class XCTestFailedError < StandardError
+
+  end
+
+  options =
+        {
+              :tries => travis_ci? ? 3 : 1,
+              :base_interval => 5,
+              #:interval => 5,
+              :on_retry => on_retry
+        }
+
+  Retriable.retriable(options) do
+    exit_code = do_system(cmd,
+                          {:pass_msg => 'XCTests passed',
+                           :fail_msg => 'XCTests failed',
+                           :exit_on_nonzero_status => false})
+    unless exit_code == 0
+      # Hunting for the exit code that indicates the simulator failed to launch.
+      log_fail "XCTest exited '#{exit_code}'"
+      raise XCTestFailedError, 'XCTest failed.'
+    end
+  end
 end