+++ CHANGELOG.md
@@ -1,4 +1,12 @@
+### 0.13.0
+
+* Migrate objc code from calabash-js submodule #123
+* Scroll gestures don't scroll past content size #120 @idcuesta
+* Fix an edge-case crash in dump route #119
+
 ### 0.12.3
 
-* Fix jsonify view crash on uislider #109
+* Add app target to xctest bundle #111
+* Fix memory leak in LPVersionRoute a0d565b
+* Replace jsonifyView: with dictionaryByEncodingView: #113 - fixes UISlider crash